refactor: use renderAsync for RNTL
