+++ packages/measure/src/index.ts
@@ -2,7 +2,7 @@ export { configure, resetToDefaults } from './config';
 export { measureRenders, measurePerformance } from './measure-renders';
 export { measureFunction } from './measure-function';
 export { measureAsyncFunction } from './measure-async-function';
-export type { MeasureRendersOptions } from './measure-renders';
+export type { MeasureRendersOptions } from './measure-renders-common';
 export type { MeasureFunctionOptions } from './measure-function';
 export type { MeasureAsyncFunctionOptions } from './measure-async-function';
 export type { MeasureType, MeasureResults } from './types';
